% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_over_steps.R
\name{apply_over_steps}
\alias{apply_over_steps}
\title{Get z scores over steps}
\usage{
apply_over_steps(a_role)
}
\arguments{
\item{a_role}{a role model}
}
\value{
data frame of step + z score
}
\description{
Get z scores over steps
}
