% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteZscores.R
\name{meteZscore}
\alias{meteZscore}
\title{meteZscore}
\usage{
meteZscore(packaged_timestep, nrep = 999, return.sim = FALSE, ...)
}
\arguments{
\item{packaged_timestep}{result of package_timestep}

\item{nrep}{n reps}

\item{return.sim}{keep sims? see meteR}

\item{...}{addtl pars passed to meteR}
}
\value{
z score
}
\description{
largely copied from meteR, calculates a Z score for obs SAD loglik compared to METE sads loglik
}
\examples{
row_from_modelSteps <- c(1, 1, 1, 8, 10, 10)
row_df <- package_timestep(row_from_modelSteps)
meteZscore(row_df)
}
