% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_role_objects.R
\name{package_timestep}
\alias{package_timestep}
\title{package timestep}
\usage{
package_timestep(row_from_modelSteps)
}
\arguments{
\item{row_from_modelSteps}{row from the modelSteps slot of a roleR object}
}
\value{
row tabled and tallied
}
\description{
package timestep
}
\examples{

row_from_modelSteps <- c(1, 1, 1, 8, 10, 10)
package_timestep(row_from_modelSteps)

}
